Change comment character
It is helpful to have commit messages include the issue number at the
beginning of a commit title. The form `#<issue-number>` does not work
if the commit is created through an editor with default settings. A line
with this prefix gets ignored by Git. Change the `core.commentChar`
value to `":"` in order to prevent this behavior.